// src/components/Forms/Booking/tipo-demanda/interfaces.ts
export interface Autor {
  id: string;
  nome: string;
  documento: string;
  email: string;
  telefone: string;
}

export interface Reu {
  id: string;
  nome: string;
  documento: string;
}

export interface Pedido {
  id: string;
  descricao: string;
  valor: string;
}

export interface CamposAuditadosCadastroJuridico {
  data_audiencia: boolean | undefined;
  numero_reserva: boolean | undefined;
  lista_autores: boolean | undefined;
}

export interface CamposAuditadosProcon {
  data_defesa: boolean | undefined;
  data_audiencia: boolean | undefined;
  numero_reserva: boolean | undefined;
  lista_autores: boolean | undefined;
}

export interface DemandaCadastroJuridico {
  pk?: string;
  created_by?: string;
  usuario_atuando?: string;
  data_carimbo?: string;
  due_date?: string;
  desdobramento?: boolean;
  cliente?: string;
  processo?: string;
  status?: string;
  perfil_demanda?: string;
  data_distribuicao?: string;
  data_distribuicao_desdobramento?: string;
  data_audiencia?: string;
  data_citacao?: string;
  acao?: string;
  tipo_acao?: string;
  tipo_acao2?: string;
  prazo_liminar?: string;
  conteudo_liminar?: string;
  prazo_tipo?: string;
  prazo_contestacao?: string;
  arquivos?: ArquivoDemanda[];
  orgao_tribunal?: string;
  uf?: string;
  cidade?: string;
  foro?: string;
  vara?: string;
  tipo_demanda?: string;
  status_demanda?: string;
  tipificacao?: string;
  processo_originario?: string;
  nome_desdobramento?: string;
  procedimento?: string;
  resumo_processo?: string;
  objeto?: string;
  pedido?: string;
  valor_pedido?: string;
  valor_causa?: string;
  causa_raiz?: string;
  sub_causa_raiz?: string;
  lista_pedidos?: Pedido[];
  lista_autores?: Autor[];
  lista_reus?: Reu[];
  numero_reserva?: string;
  observacao?: {
    mensagem: string;
    criada_em: string;
    criada_por: string;
  }[];

  campos_auditados?: CamposAuditadosCadastroJuridico;
}
// ...
export interface FormFile {
  id: string;
  info: {
    name: string;
    size: number;
    sizeFormatted: string;
    unit: Unit;
    url: string; 
    tipo?: string;
    s3Key: string;
    s3Bucket: string;
  };
  status: {
    success: boolean | undefined;
    progress: number;
    message: string | undefined;
    isDeleting?: boolean;
  };
  allow: {
    delete: boolean;
    download: boolean;
    retryUpload: boolean;
    link: boolean;
  };
  dropzoneFile: File | FileRejection | undefined;
}
// ...
export type TipoJustificativaInfo = 'enviar-excecao-cliente' | 
'enviar-esteira-cliente' | 'enviar-esteira-oito' | 'enviar-auditoria-oito' |
'enviar-email' | 'substituir-endereco-email' | 'inativar-demanda' |
'enviar-excecao-oito' | 'reprovar-demanda' | '';

export interface JustificativaInfo {
  tipo: TipoJustificativaInfo;
  title: Partial<{
    text: string;
  }>,
  justificativa: Partial<{
    isValid: boolean;
    value: string;
  }>
}